Load metadata from program comments.
