+++ adli.py
@@ -6,6 +6,16 @@ import argparse
 import uuid
 from injector.ProgramProcessor import ProgramProcessor
 
+'''
+{
+    "type": "adli_metadata",
+    "name": "ADLI",
+    "description": "A tool to inject diagnostic logs into a python program.",
+    "version": "0.0",
+    "language": "python"
+}
+'''
+
 def verify_python_compatibility():
     if not hasattr(ast, 'unparse'):
         raise RuntimeError("This program requires Python 3.9+ (ast.unparse not available)")
@@ -65,7 +75,8 @@ def main(argv):
             print(f"Error reading system info file: {str(e)}", file=sys.stderr)
             return -1
     else:
-        sysinfo = None
+        sysinfo = {}
+
     # If no unique id was provided, generate one
     if (uid == None):
         uid = str(uuid.uuid4())
